Billy's Factorization script
Crack.py implements a 3 hidden layers neural network, trained on 50,000
randomly selected 32bits semi-primes to find the 16 bit smaller factor.
The result is evaluated on a different set of 1000 semi-primes. The
model is saved as model.ckpt and the outputs of two different runs of
the script are directed into the files, run1.out and run2.out